made the board size button panel its own view and refactored some naming of packages
