+++ src/main/interface_adapter/boardSizeSelection/BoardSizeViewModel.java
@@ -1,6 +1,6 @@
-package main.interface_adapter.characterSelectionScreen;
+package main.interface_adapter.boardSizeSelection;
 
-import main.use_case.BoardSizeSelection.BoardSizeSelection.BoardSize;
+import main.use_case.boardSizeSelection.BoardSizeSelection.BoardSize;
 
 /**
  * View model for board size selection.
