+++ src/main/use_case/boardSizeSelection/BoardSizeInputBoundary.java
@@ -1,6 +1,6 @@
-package main.use_case.BoardSizeSelection;
+package main.use_case.boardSizeSelection;
 
-import main.use_case.BoardSizeSelection.BoardSizeSelection.BoardSize;
+import main.use_case.boardSizeSelection.BoardSizeSelection.BoardSize;
 
 /**
  * Input boundary interface for board size selection use case.
