+++ src/main/use_case/boardSizeSelection/BoardSizeOutputBoundary.java
@@ -1,6 +1,6 @@
-package main.use_case.BoardSizeSelection;
+package main.use_case.boardSizeSelection;
 
-import main.use_case.BoardSizeSelection.BoardSizeSelection.BoardSize;
+import main.use_case.boardSizeSelection.BoardSizeSelection.BoardSize;
 
 /**
  * Output boundary interface for board size selection presenter.
